refactor(AutocompleteE): Consolidate hooks
Hooks need to interact with each other so it makes sense to consolidate.
